Load demo tests using sys.executable
Previously, the string "python" was being used
in the subprocess call. However, this did not always
do the right thing when, for example, the demo tests
are being called within a virutal environment using
pytest. By using sys.executable, we use the same
python interpreter used to run pytest and/or the test_demos.py
file.