Annotate all pub struct with #[non_exhaustive]
This prevents code outside of the crate from
constructing the given structs using a StructExpression
(i.e., MyStruct { ... }). I have tested and confirmed
that, for example, I cannot instatiate a struct from
the fastsim-cli crate. This was deemed to be a little
less convoluted than using a PhantomData typed method
which essentially does the same thing.